requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
argon2-cffi==25.1.0
    # via backend (pyproject.toml)
argon2-cffi-bindings==25.1.0
    # via argon2-cffi
asyncpg==0.30.0
    # via backend (pyproject.toml)
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
    #   sentry-sdk
cffi==2.0.0
    # via argon2-cffi-bindings
click==8.3.1
    # via
    #   rich-toolkit
    #   typer
    #   uvicorn
dnspython==2.8.0
    # via email-validator
email-validator==2.3.0
    # via
    #   fastapi
    #   pydantic
fastapi==0.121.3
    # via backend (pyproject.toml)
fastapi-cli==0.0.16
    # via fastapi
fastapi-cloud-cli==0.5.1
    # via fastapi-cli
fastar==0.6.0
    # via fastapi-cloud-cli
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via
    #   fastapi
    #   fastapi-cloud-cli
idna==3.11
    # via
    #   anyio
    #   email-validator
    #   httpx
itsdangerous==2.2.0
    # via fastapi
jinja2==3.1.6
    # via fastapi
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
orjson==3.11.4
    # via fastapi
pycparser==2.23
    # via cffi
pydantic==2.12.4
    # via
    #   fastapi
    #   fastapi-cloud-cli
    #   pydantic-extra-types
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-extra-types==2.10.6
    # via fastapi
pydantic-settings==2.12.0
    # via fastapi
pygments==2.19.2
    # via rich
python-dotenv==1.2.1
    # via
    #   backend (pyproject.toml)
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.20
    # via fastapi
pyyaml==6.0.3
    # via
    #   fastapi
    #   uvicorn
rich==14.2.0
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.16.0
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
rignore==0.7.6
    # via fastapi-cloud-cli
sentry-sdk==2.45.0
    # via fastapi-cloud-cli
shellingham==1.5.4
    # via typer
sniffio==1.3.1
    # via anyio
starlette==0.50.0
    # via fastapi
typer==0.20.0
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
typing-extensions==4.15.0
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   pydantic-extra-types
    #   rich-toolkit
    #   starlette
    #   typer
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
ujson==5.11.0
    # via fastapi
urllib3==2.5.0
    # via sentry-sdk
uvicorn==0.38.0
    # via
    #   backend (pyproject.toml)
    #   fastapi
    #   fastapi-cli
    #   fastapi-cloud-cli
uvloop==0.22.1
    # via uvicorn
watchfiles==1.1.1
    # via uvicorn
websockets==15.0.1
    # via uvicorn
